MetadataExtraction/requirements.txt
SQLAlchemy==1.3.23
Flask_Cors==3.0.10
SQLAlchemy_Utils==0.36.8
Flask==1.1.2
Flask_SQLAlchemy==2.4.4
Pillow==8.1.0
mysqlclient==2.0.3